requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
cryptography==46.0.3
dnspython==2.7.0
email-validator==2.3.0
Faker==37.6.0
Flask==3.1.2
flask-cors==6.0.1
Flask-Dance==7.1.0
Flask-Login==0.6.3
flask-marshmallow==1.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
freezegun==1.5.5
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.1
marshmallow-sqlalchemy==1.4.2
oauthlib==3.3.1
packaging==25.0
passlib==1.7.4
playwright==1.55.0
pluggy==1.6.0
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.23
pyee==13.0.0
Pygments==2.19.2
pyjokes==0.8.3
pyOpenSSL==19.0.0
pytest==8.4.2
pytest-base-url==2.1.0
pytest-freezegun==0.4.2
pytest-playwright==0.7.0
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
requests==2.32.5
requests-oauthlib==2.0.0
setuptools==80.9.0
six==1.17.0
soupsieve==2.8
SQLAlchemy==2.0.43
text-unidecode==1.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
URLObject==3.0.0
Werkzeug==3.1.3
WTForms==3.2.1